pyproject.toml
[project]
name = "brave-search-mcp"
version = "0.1.0"
description = "MCP server for integrating Brave Search into AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.0"
]